broke apart game importer files into smaller extensions in preparation to strange everything through the import queue
